Checked table scripts for data types, length, nulls, etc.

+++ SoS_Build/SoS_Build/CreateTable_luPhoneTypes.sql
@@ -1,7 +1,7 @@
 USE [SoS_Build]
 GO
 
-/****** Object:  Table [dbo].[luPhoneTypes]    Script Date: 3/13/2018 6:02:11 PM ******/
+/****** Object:  Table [dbo].[luPhoneTypes]    Script Date: 3/25/2018 9:02:58 AM ******/
 SET ANSI_NULLS ON
 GO
 
@@ -9,7 +9,7 @@ SET QUOTED_IDENTIFIER ON
 GO
 
 CREATE TABLE [dbo].[luPhoneTypes](
-	[PhoneTypeID] [int] IDENTITY(1,1) NOT NULL,
+	[PhoneTypeID] [tinyint] IDENTITY(1,1) NOT NULL,
 	[PhoneType] [nvarchar](10) NOT NULL,
  CONSTRAINT [PK_luPhoneTypes] PRIMARY KEY CLUSTERED 
 (
@@ -17,5 +17,4 @@ CREATE TABLE [dbo].[luPhoneTypes](
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
 ) ON [PRIMARY]
 GO
-
